+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Karafka framework changelog
 
 ## 2.1.0 (Unreleased)
+- **[Feature]** Introduce collective Virtual Partitions offset management
+- [Improvement] Always use Virtual offset management for Pro ActiveJobs.
 - [Improvement] Do not attempt to mark offsets on already revoked partitions.
+- [Improvement] Make sure, that VP components are not injected into non VP strategies.
 
 ## 2.0.41 (2023-14-19)
 - **[Feature]** Provide `Karafka::Pro::Iterator` for anonymous topic/partitions iterations and messages lookups (#1389 and #1427).